requirements.txt
gradio~=5.22.0
python-dotenv~=1.0.1
openai~=1.68.2
uvicorn
fastapi